I2C communication code now stores reads into byte instead of uint16_t.

+++ src/flightcode/i2c.h
@@ -3,14 +3,14 @@
 
 #include <Wire.h>
 
-int sendI2C(int device, uint8_t address, uint8_t val) {
+int sendI2C(int device, byte address, byte val) {
    Wire.beginTransmission(device);
    Wire.send(address);
    Wire.send(val);
    return Wire.endTransmission();
 }
 
-int readI2C(int device, uint8_t address, int num, uint16_t buff[]) {
+int readI2C(int device, byte address, int num, byte buff[]) {
     int eCode;
 
     Wire.beginTransmission(device);
